fix!: loading with out remount
previously loading works by mount a loading animation or a null. with
this commit we change the behaviour of loading by utilizing
`display:none`. the approach is not using a `hidden` by tailwind because
JSDom has bug with css changes after render.

the other changes is moving loading to be part of authentication
features. this movement because no other component use loading other
than Protected Page.